Try it out: at the root of our project, create a new foods.js file. Then, copy the old foods code from play.js, delete it, and paste it in foods.js.
Modules & module.exports

Now, as you're probably expecting, we're going to require foods.js from play.js. But there is one really important difference between the way require works in PHP versus Node. In PHP, when you require a file, you magically have access to all functions, classes or variables inside. But in Node, that's not true: you need to explicitly export a value from the required file.

How? By saying module.exports = foods.

By the way, in JavaScript, each file is known as a module and is executed in isolation. Nothing is returned from this file other than what we export.

Ok! Back in play.js, we can now say, const foods = require('./foods').
Before I say more, try it! Head back to your terminal and re-run the code:

node play.js
It still works! Back in the code, notice that we do not need the .js on the end of the filename.

We can add it - that would totally work. But if it's not there, Node knows to look for foods.js.

Also, that ./ before foods is no accident: that's super important. When a path starts with a ., Node knows to look for that file relative to this file. If we just said foods without the ./, well, that means something very different. More on that later.