Using require in the Browser
So the require() function is an awesome feature. The question is: can we use this in the browser JavaScript world? Because if we could, just imagine how easy it would be to organize our JavaScript!

And yes! We can totally get it working... otherwise this would be a really short tutorial.

But first, a little bit of cleanup: open app/Resources/views/lift/index.html.twig.

This is the template for the main page. And at the end of the last tutorial, we used a library called Babel to 'transpile' our source RepLogApp.js into a dist/RepLogApp.js file. We are going to use Babel again, but for now, delete the dist file. Then, point our page back to the source RepLogApp.js. This file contained two classes: RepLogApp and, near the bottom, another called Helper.

Hey! For organization, let's move this class into its own file. In that same directory, create a new RepLogAppHelper.js file. I'll add the 'use strict'; on top and then paste the class.

At the bottom, add module.exports = Helper.

You can export anything you want from a module - a value, like we did earlier - a class, a function, an object - whatever!

Back in RepLogApp.js, now that the Helper class is gone... well, we're not going to have a good time. On line 10, PhpStorm is giving me a cryptic error: element is not exported... a funny way of saying 'Variable undefined' !