Add const Helper = require('./RepLogAppHelper')
My editor is happy!

Based on what we saw in Node... this should just work! Back in my browser, I'll open the console and then, refresh. Yep, we somehow knew life wouldn't be so simple. We get an error:

require is not defined
Here's the deal: the require() function does not work in any browser. And it's not that browsers are behind... it's just not possible to make it work!

Think about it: in PHP, when we use the require statement, we're reading a file from our file system... which is basically instant. But on the web, a browser would need to go download that file. Imagine if we waited while it downloaded this file... then this file required 5 other files.... so we waited for those... then those files required 10 other files... so we finally decide to go have lunch while the web page loads. It just doesn't work!

Enter Webpack.