Since we programmers are famously lazy, we don't usually run Webpack with all these cumbersome command-line options. Nope, we put them into a configuration file. At the root of your project, create a new webpack.config.js file. If you make a file with exactly this name, Webpack will automatically find it.
And what does this file look like? Oh, you're going to love this: we use, yep module.exports to export a configuration object!
Inside, that config needs two major things. First, an entry key set to ./web/assets/js/RepLogApp.js. Entry is Webpack's name for your input file: the one file it will load. Second, of course, we need to tell Webpack where to create the output. Set this to a hash with path set to the directory: __dirname plus /web/build. That __dirname is a special variable available inside of Node that points to the directory of this file. Then, add a filename key set to rep_log.js.

Now, running Webpack is truly a lazy experience... we should probably lift our cat a few times for exercise. Just run:

./node_modules/.bin/webpack
And it works exactly like before.