The downside of using something like Webpack - or any tool that builds your assets - is that every time you make a change, you have to re-run webpack. That's lame.
Fortunately, like many tools, Webpack has a 'watch mode', and it just works. Right now, if I click delete, it says, 'Delete this log?'

Re-run webpack, but this time with a --watch flag:
sudo ./node_modules/.bin/webpack --watch

It still builds the assets... but then waits. It's now constantly watching all your files for changes. Creepy.

Put it to the test: in RepLogApp.js, add a few more question marks to the delete modal for emphasis and save.
Refresh the page and.... we instantly see the change.

This is nothing earth-shattering... it's just a tool we need... and Webpack provides it without any setup. In the terminal, you can even see that it re-dumped our file.

There is one caveat with watch: and we'll see it a few times. If you make a change to the webpack.config.js file itself, you will need to manually restart webpack. Webpack watch does not detect changes to itself.