The Classic Problem with JS Dependencies
Ok, it's time to do something completely different... and unlock a massively powerful new tool. I'm going to say that a lot in this tutorial... and I mean it every time!

Right now, RepLogApp only works because jQuery is available globally:

 217 lines  web/assets/js/RepLogApp.js
... lines 1 - 4
(function(window, {dollar_sign}, Routing, swal) {
... lines 6 - 215
})(window, jQuery, Routing, swal);
We have a self-executing function, which expects jQuery to be a global variable by the time this file is loaded. That works because, in our base layout - app/Resources/views/base.html.twig - we include jQuery:

 110 lines  app/Resources/views/base.html.twig
... lines 1 - 98
{% block javascripts %}
    <script src='https://code.jquery.com/jquery-3.1.1.min.js' integrity='sha256-hVVnYaiADRTO2PzUGmuLJr8BLUSjGIZsDYGmIJLv2b8=' crossorigin='anonymous'></script>
... lines 101 - 105
{% endblock %}
... lines 107 - 110
This gives us two global variables - {dollar_sign} and jQuery - which we use everywhere else.

And this is the big, huge, classic, horrifying problem we want to fix: when we write JavaScript, our code is not self-contained. Nope, we need to very carefully make sure to remember to include any dependent script tags before loading our JavaScript. If you forget to add a script tag or add it in the wrong order... bam! Your user sees an error. And you see your face hitting your palm.

This is a terrible and unprofessional way to live. But with Webpack, we can finally fix this. Instead of crossing your fingers and hoping that jQuery is included on this page, in RepLogApp.js, we will require it.