At the top of RepLogApp.js, add const {dollar_sign} = require('jquery') For simplicity, we can remove the {dollar_sign} from the self-executing function and the jQuery argument at the bottom.
Now, we require jQuery from node_modules and assign it to the {dollar_sign} variable. Whenever we reference {dollar_sign} in this file, that required value is used. We are no longer dependent on whether or not a global {dollar_sign} or jQuery variable exists.

What Happens when you Require a Module
There are two very important things I want to point out. First, we just said require('jquery').
We now know that because this does not start with ., Node knows to look for a core library - like it did with path - or to look inside node_modules/, which is what happens this time. That's perfect!

The second very, very, very important thing is that jQuery acts differently, depending on whether or not it's being included with a traditional script tag - like in our layout - or if it's being required by a module system. This is a very common thing you'll see.

Let me show you: I'll hold command to click into the jquery module itself. Yea, it's a bit hard to read. But, on top, it checks to see if typeof module === 'object', and typeof module.exports === 'object'. Basically jQuery is checking whether or not it is being used inside of a module environment, called a commonjs environment.

If it is, it does not create global jQuery or $ variables. Nope, instead it uses module.exports to export the jQuery function. This is really important: it means that when you require jquery, it returns the jQuery function... but does not actually create a global variable. This is a pivotal difference between classic JavaScript and modern JavaScript: instead of creating and using global variables, we require modules and export things from these modules.

Phew! Let's try this out already! In my terminal, I'll check out my watch tab. Yep, it's still working, and should have already detected our new require() and dumped the new file.

Back in the browser, refresh! Yes! Everything still works! It's not super obvious yet, but RepLogApp is now using the required jQuery instead of the global script tag.

But, I can't remove that script tag from the base layout yet... because we have other pages that depend on it. And yes, that means that - until we fix this - our users are downloading jQuery twice: once in the base layout and again when they download rep_log.js.

Open up that file: web/build/rep_log.js. Yep, most of this file is now jQuery.

Now that we have this super power, let's repeat it with SweetAlert... and discover one lingering issue.