Find your terminal and run: yarn add sweetalert2 --dev Hey! We can install and require third-party libraries. That's amazing! And it makes our code more reliable: we do not need to hope that someone remembered to include jQuery on this page.
RepLogApp also depends on two other vendor libraries: Routing from FOSJsRoutingBundle and SweetAlert2.

Let's save Routing for later and skip straight to SweetAlert.

It's the exact same situation as jQuery, except that - this time - the script tag is included on this specific page, not in the layout. Open app/Resources/views/lift/index.html.twig. At the bottom, yep! There is the script tag for SweetAlert.

We do not want that anymore!

First, we need to download SweetAlert via Yarn. If you're not sure about the name of a package, check out https://npms.io/. Search for 'SweetAlert2'. The first looks like a good match and... when we click it... yep! This is the exact library we're using.