Require sweetalert2
Perfect! Back in RepLogApp, add const swal = require('sweetalert2'). Then, remove the extra swal from the top of the self-executing function and on the bottom: there's no need for that complexity.

And guess what! SweetAlert is only used by RepLogApp... so we can safely remove the script tag from the template. Woohoo!

Try it: head back to LiftStuff and refresh. Yes! It still loads!

But when we hit 'Delete'... woh... something isn't right: it looks like I designed this... which you do not want. And... I can't hit 'Cancel'. What's going on here?