Exporting RepLogApp Since RepLogApp is being loaded by webpack. it's already a module. So instead of using window.RepLogApp, let's export a value properly: module.exports = RepLogApp.
Now, if anything requires this file, they will get the RepLogApp class. And, we are no longer modifying anything in the global scope.

And as soon as we try that, our app is super broken! Thanks Ryan!

RepLogApp is not defined
Yes! This makes sense: in our template, we're still trying to reference the now,

non-existent - global variable RepLogApp:
 66 lines  app/Resources/views/lift/index.html.twig
... lines 1 - 53
{% block javascripts %}
... lines 55 - 58
  <script>
    {dollar_sign}(document).ready(function() {
... line 61
      var repLogApp = new RepLogApp({dollar_sign}wrapper);
    });
  </script>
{% endblock %}