webpack.config.js & Multiple entry
Next, open webpack.config.js. We have a problem: we really only have the ability to configure one entry file... which creates one output file.

How can we tell Webpack to also look at login.js and build a second file? The secret is that entry can actually contain multiple things.

Copy the original path and set entry to {}. For the first entry, put rep_log as the key set to the copied path. Then, repeat that for a new entry called login that points to our source login.js file. Woohoo! Webpack will now read these two entry files. To have it output two built files, we need to change the filename. In place of rep_log, use a special placeholder: [name].js. The name is referring to the keys I created under entry: rep_log and login. These could have been anything. For consistency, they match the source filenames, but really, you can choose whatever you want. The only significance of these keys is that they become the name of the final, dumped file.

You guys know the drill: we just made a change to webpack.config.js. And that means it's time to stop the watch script with Control+C and restart it.
Ah... yes! It dumped login.js. There it is, sitting beautifully inside build/.

In login.html.twig, make sure to update the script tag to point to build/login.js. Then, head over to your browser and refresh the rep log page first. It still works. Go to /login... and it works! This was another huge step! We now have the power to create as many different distinct JavaScript applications as we need.