Now that layout.js is being processed through webpack, we can remove our self-executing function and replace the $ with const $ = require('jquery') We rock! But this time... there's a surprise. When we refresh, we do not get warm, fuzzy feelings: we get a cold and not-so-fuzzy error:
$.tooltip() is not a function
Interesting!

 8 lines  web/assets/js/layout.js
... lines 1 - 4
{dollar_sign}(document).ready(function() {
  {dollar_sign}('[data-toggle='tooltip']').tooltip();
});
This $.tooltip() function should be coming from Bootstrap... which is included in our base layout:

 110 lines  app/Resources/views/base.html.twig
... lines 1 - 98
{% block javascripts %}
... line 100
  <script src='https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js' integrity='sha384-Tc5IQib027qvyjSMfHjOMaLkfuWVxZxUPnCJA7l2mCWNIpG9mGCD8wGNIcPD7Txa' crossorigin='anonymous'></script>
... lines 102 - 105
{% endblock %}
... lines 107 - 110
So... why isn't this working? Well... think about it: Bootstrap's job is to modify jQuery and add more functions to it, like 'tooltip'. But, in layout.js, we require a different, fresh, unmodified jQuery and then try to call .tooltip() on it!

That doesn't work because - when the Bootstrap code is loaded - it modifies the jQuery that's loaded in our layout: it modifies the global jQuery variable. In layout.js, we're not using the global jQuery variable: we're requiring our own jQuery... which has not been modified by Bootstrap.

The fix? We already know it: we need to once again handle all our dependencies inside layout.js. I mean, if we need bootstrap functions inside of this file, then we need to require bootstrap here.

Let's do it!