Using a Loader Globally
This is great! Except that I do not want to have to add this to every require statement! Thankfully, Webpack also has a global way to apply loaders.

Remove the inline loader syntax. And open webpack.config.js. Add a new module key set to {} and a sub-key called rules set to an array.
Here's the deal: each rule will contain a filename regular expression and a loader that should be applied whenever a file matches that. Add {} for this first loader with a test key. We want to apply the loader to all files that end in .js

That's what this regular expression matches. Below this, add use, with loader: 'babel-loader'.

That is it! Now, every .js file will go through Babel! Woohoo!

Restart the webpack script:
./node_modules/.bin/webpack --watch

And check out the built rep_log.js file. Try to find RepLogApp again. Yep, the same, beautiful, transpiled code.