Now, you may or may not have noticed it, but that build took a bit longer than normal: it took almost five seconds. That makes sense... our code is now going through Babel.
But actually, most of that time is because every file is being processed through Babel... including third-party files like jQuery and Bootstrap! That's overkill.

Skipping node_modules Files

Most of the time, we expect third-party modules to already be compiled into older JavaScript. So, we don't need to re-transpile them. That may not always be the case, so just be cautious.

How can we skip third-party files? Add an exclude for /node_modules/

And just by making this small change, when we re-run Webpack, it's way faster:
./node_modules/.bin/webpack --watch

Instead of four and a half seconds, we're back under two seconds.