Configuring publicPath
The image was not there! And my console had a 404 for the image! What's up?!

Ok, the final CSS from Webpack changed the background-image to point to the new filename. When opened that in a new tab, the filename was right, but it's missing the build/ directory prefix!

Webpack was almost doing everything correctly: it moved the file into build/ and even updated the CSS to point to that filename.

Open webpack.config.js. Yes, we did tell Webpack to put everything into web/build. But, Webpack didn't actually know what the public path is to files in this directory. I mean, it didn't know that web/ is our document root, and so it didn't know that these files are accessible via build/ then the filename. Nope, this is something we need to tell Webpack, so that it can create the correct paths.

How? Under output, set publicPath to /build/

Find you terminal and restart Webpack:

./node_modules/.bin/webpack --watch
Everything looks the same here... but when we refresh and open the menu... there it is! Our little icon. The background-image now point to /build/ the filename.

Guys, this is another monumental step forward! Now, as long as we correctly reference image paths in CSS, Webpack will make sure those images are available in build/ and that their paths in the final CSS are correct. We focus on our source files, and Webpack takes care of the rest.

Even better, if you make a mistake - like a typo - you'll actually see a Webpack build error. There's no way to accidentally have a broken link.