Next, in layout.js, you guys know what to do - add require('font-awesome/'). To get the correct path, once again, open up node_modules/, find font-awesome/, and yes! We can say css/font-awesome.css. And without doing anything else, refresh! It works instantly! This little icon in the menu is from FontAwesome, and so are a few other icons.
The font-loader Hashed Filenames

Close node_modules/ and look inside web/build/. Wow! It's getting crowded! We have fonts from Bootstrap, fonts from FontAwesome and one image.

And, though it's not really important, but we can't quickly see which files are which... because they all have crazy names.

By default, the filenames are a unique hash based on the contents of the file. That's great, because it means that if you reference two files that have the same name... but usually live in different directories, when they're copied to build/, they won't collide. It's also built-in cache busting!