Controlling file-loader Hashes
Head back to the Webpack site and click back on 'Loading Images'. There's a link to the documentation for the file-loader(https://webpack.js.org/loaders/file-loader/).

It shows how to use the file-loader using the inline syntax, which is one of the reasons I wanted you to see it. It looks like there's an option called name... which should let us control the name of those output files. And, just like with Webpack's output, it has some special placeholders, like [name], [extension] and [hash].

Let's play with this! In webpack.config.js, remove the simple file-loader and use the expanded syntax from earlier. That means, add loader: 'file-loader' and options.

Pass this the name option, set to [name] - that's the original name of the file - -[hash] because we do still want the hash in the name. But shorten it by adding :6. This will only use the first six characters of the hash, which is probably unique enough - these long filenames were bumming me out. Finish with .[ext]. This is totally not needed - I'm just being anal about how my files are named. Copy all of this and repeat it for the fonts.

To see this in action, in your terminal, stop webpack, and then clear out the build/ directory:
rm -rf web/build/*
Ok, come back Webpack!

./node_modules/.bin/webpack --watch
Yes! Now, the filenames are way prettier. And I feel better!

Next! Let's get crazy and move our front-end assets... out of the public directory...