I just read this awesome article about how cool Sass is. So now, I'm bored with my old, main.css file. Oof. Ancient! Rename this to main.scss. Trendy! Exciting!
Now that this is a Sass file - oooOOOooo - I'll re-work some syntax. Instead of having this extra :hover and :focus section, we can leverage Sass syntax and say &:hover, &:focus, followed by the CSS.

That should, ultimately, dump the same CSS. But now, we have the power of Sass!

Since we renamed the file, in layout.js, we need to change the require to be main.scss.

Now, what do you think will happen when Webpack sees this import? You can probably guess: it's not going to like it.

Find your webpack terminal. Oh boy... oof, I was right. We have a very familiar error:

Module parse failed. Unexpected token. You may need an appropriate loader to handle this file.
We do have a loader for files ending in .css... but not .scss. Fixing this? Oh, it's so nice.