This loader exists only to solve this problem. Once it's installed, find webpack.config.js and, right before sass-loader, add resolve-url-loader. But, for this to work, on sass-loader, add ?sourceMap.
We're going to talk more about source maps in a few minutes. But internally, the resolve-url-loader needs them so it can do its job.

I know. It's a little crazy. But! If we go back and restart Webpack...
./node_modules/.bin/webpack --watch

It works! And the page comes back to life!

Hey! We've got Bootstrap via Sass! Now, let's tweak some stuff!